Safari-611.1.10/LayoutTests/inspector/audit/basic-object-expected.txt

Testing the basic functionality of audits.


== Running test suite: Audit.Basic
-- Running test case: Audit.Basic.Object.Pass
Testing value `{"level":"pass"}`...
PASS: Result should be "pass".

-- Running test case: Audit.Basic.Object.Warn
Testing value `{"level":"warn"}`...
PASS: Result should be "warn".

-- Running test case: Audit.Basic.Object.Fail
Testing value `{"level":"fail"}`...
PASS: Result should be "fail".

-- Running test case: Audit.Basic.Object.Error
Testing value `{"level":"error"}`...
PASS: Result should be "error".

-- Running test case: Audit.Basic.Object.Unsupported
Testing value `{"level":"unsupported"}`...
PASS: Result should be "unsupported".
